requirements.txt
fusepy==2.0.4
pyyaml==5.1
b2-sdk-python>=1.8.0,<2.0.0
intervaltree==3.1.0